# Conflicts:
#	kalium

+++ app/src/main/kotlin/com/wire/android/ui/home/conversationslist/model/EventType.kt
@@ -1,7 +1,7 @@
 package com.wire.android.ui.home.conversationslist.model
 
 sealed class EventType {
-    data class UnreadMessage(val unreadMessageCount: Int) : EventType()
+    data class UnreadMessage(val unreadMessageCount: Long) : EventType()
     object UnreadMention : EventType()
     object UnreadReply : EventType()
     object MissedCall : EventType()
